refactor(build): Centralize Kover and Dokka configurations
This commit refactors the build logic by creating dedicated convention plugins for Kover (code coverage) and Dokka (documentation).

Key changes:
- Created `KoverConventionPlugin` to apply and configure Kover for each module.
- Moved Dokka configuration logic from `DokkaConventionPlugin` into a reusable `configureDokka` function.
- Removed duplicated Kover and Dokka plugin applications and configurations from individual module `build.gradle.kts` files.
- Centralized Kover and Dokka aggregation logic in the `RootConventionPlugin`.
- Updated Android and KMP convention plugins to apply the new `meshtastic.kover` and `meshtastic.dokka` plugins.
- Updated copyright years across various build files.

This centralization simplifies module build scripts, reduces redundancy, and makes the build configuration more maintainable.
