+++ build-logic/convention/src/main/kotlin/AndroidLibraryConventionPlugin.kt
@@ -33,6 +33,7 @@ class AndroidLibraryConventionPlugin : Plugin<Project> {
             apply(plugin = "meshtastic.detekt")
             apply(plugin = "meshtastic.spotless")
             apply(plugin = "meshtastic.dokka")
+            apply(plugin = "meshtastic.kover")
 
             extensions.configure<LibraryExtension> {
                 configureKotlinAndroid(this)
