+++ build.gradle.kts
@@ -45,78 +45,6 @@ plugins {
 
 
 
-kover {
-    reports {
-        filters {
-            excludes {
-                // Exclude generated classes
-                classes("*_Impl")
-                classes("*Binding")
-                classes("*Factory")
-                classes("*.BuildConfig")
-                classes("*.R")
-                classes("*.R$*")
-
-                // Exclude UI components
-                annotatedBy("*Preview")
-
-                // Exclude declarations
-                annotatedBy(
-                    "*.HiltAndroidApp",
-                    "*.AndroidEntryPoint",
-                    "*.Module",
-                    "*.Provides",
-                    "*.Binds",
-                    "*.Composable",
-                )
-
-                // Suppress generated code
-                packages("hilt_aggregated_deps")
-                packages("org.meshtastic.core.strings")
-            }
-        }
-    }
-}
-
 dependencies {
-    kover(projects.app)
-
-    kover(projects.core.analytics)
-    kover(projects.core.common)
-    kover(projects.core.data)
-    kover(projects.core.database)
-    kover(projects.core.datastore)
-    kover(projects.core.di)
-    kover(projects.core.model)
-    kover(projects.core.navigation)
-    kover(projects.core.network)
-    kover(projects.core.prefs)
-    kover(projects.core.proto)
-    kover(projects.core.service)
-    kover(projects.core.ui)
-    kover(projects.feature.intro)
-    kover(projects.feature.messaging)
-    kover(projects.feature.map)
-    kover(projects.feature.node)
-    kover(projects.feature.settings)
-    kover(projects.feature.firmware)
-
-    // Aggregated Dokka documentation
-    subprojects.forEach { subproject ->
-        subproject.pluginManager.withPlugin("org.jetbrains.dokka") {
-            dokka(subproject)
-        }
-    }
-
     dokkaPlugin(libs.dokka.android.documentation.plugin)
-}
-
-
-// ... rest of the file ...
-
-dokka {
-    moduleName.set("Meshtastic App")
-    dokkaPublications.html {
-        suppressInheritedMembers.set(true)
-    }
 }
