+++ core/ui/build.gradle.kts
@@ -37,7 +37,6 @@ plugins {
     alias(libs.plugins.meshtastic.android.library)
     alias(libs.plugins.meshtastic.android.library.compose)
     alias(libs.plugins.meshtastic.hilt)
-    alias(libs.plugins.kover)
 }
 
 configure<LibraryExtension> { namespace = "org.meshtastic.core.ui" }
