+++ feature/firmware/build.gradle.kts
@@ -34,7 +34,6 @@ import com.android.build.api.dsl.LibraryExtension
  */
 
 plugins {
-    alias(libs.plugins.kover)
     alias(libs.plugins.meshtastic.android.library)
     alias(libs.plugins.meshtastic.android.library.compose)
     alias(libs.plugins.meshtastic.hilt)
